% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{exercise}
\alias{exercise}
\title{Open an exercise}
\usage{
exercise(chapter, fname)
}
\arguments{
\item{chapter}{The chapter}

\item{fname}{The filename}
}
\description{
\code{exercise} opens an exercise file in your RStudio scripts pane and 
saves a copy of the file in your working directory.
}
\details{
Heavily based on the code in the reportsWS package.
}
