% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_show.R
\name{layout_show}
\alias{layout_show}
\title{Show complex layouts}
\usage{
layout_show(n = 1, cex = 1, text = NULL)
}
\arguments{
\item{n}{number of figures to plot.}

\item{cex}{text size}

\item{text}{Text to print}
}
\description{
Tweak to \code{layout.show} to allow text size to be displayed.
}
