% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setnicepar.R
\name{setnicepar}
\alias{setnicepar}
\title{Sets par parameters}
\usage{
setnicepar(mar = c(3, 3, 2, 1), mgp = c(2, 0.4, 0), tck = -0.01,
  cex.axis = 0.9, las = 1, mfrow = c(1, 1), ...)
}
\arguments{
\item{mar}{Distance to edge}

\item{mgp}{Distance from label to graph}

\item{tck}{tick length}

\item{cex.axis}{Axis text size}

\item{las}{Rotate axis label}

\item{mfrow}{plots per page}

\item{...}{Additional arguments for par}
}
\description{
Set more sensible par parameters
}
\seealso{
par
}
