% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{solution}
\alias{solution}
\title{Open a solution}
\usage{
solution(number = 1)
}
\arguments{
\item{number}{The number of solution to open. Solution numbers are paired with exercises.}
}
\description{
\code{solution} opens a solution file in your RStudio scripts pane and saves a copy of the file in your working directory.
}
\details{
Heavily based on the code in the \code{reportsWS} package.
}
\seealso{
\code{\link{exercise}}
}
